ver2.0/requirements.txt
deepl==1.5.0
Flask==2.0.2
gunicorn==20.1.0
line-bot-sdk==1.20.0
selenium==3.141.0
